+++ Hive/Source/HiveLib/DataSource/SqlObjDataSource.h
@@ -31,14 +31,14 @@ public:
 
 	void populateObjects( int serverId, ServerObjectsQueue& queue ) override;
 
-	void populateTraderObjects( int characterId, ServerObjectsQueue& queue ) override;
+	void populateTraderObjects( Int64 characterId, ServerObjectsQueue& queue ) override;
 
 	bool updateObjectInventory( int serverId, Int64 objectIdent, bool byUID, const Sqf::Value& inventory ) override;
 	bool deleteObject( int serverId, Int64 objectIdent, bool byUID ) override;
 	bool updateDatestampObject( int serverId, Int64 objectIdent, bool byUID ) override;
 	bool updateVehicleMovement( int serverId, Int64 objectIdent, const Sqf::Value& worldspace, double fuel ) override;
 	bool updateVehicleStatus( int serverId, Int64 objectIdent, const Sqf::Value& hitPoints, double damage ) override;
-	bool createObject( int serverId, const string& className, double damage, int characterId, 
+	bool createObject( int serverId, const string& className, double damage, Int64 characterId, 
 		const Sqf::Value& worldSpace, const Sqf::Value& inventory, const Sqf::Value& hitPoints, double fuel, Int64 uniqueId ) override;
 	Sqf::Value fetchObjectId( int serverId, Int64 objectIdent ) override;
 private:
