+++ Hive/Source/HiveLib/HiveExtApp.cpp
@@ -399,7 +399,7 @@ Sqf::Value HiveExtApp::objectPublish( Sqf::Parameters params )
 {
 	string className = boost::get<string>(params.at(1));
 	double damage = Sqf::GetDouble(params.at(2));
-	int characterId = Sqf::GetIntAny(params.at(3));
+	Int64 characterId = Sqf::GetIntAny(params.at(3));
 	Sqf::Value worldSpace = boost::get<Sqf::Parameters>(params.at(4));
 	Sqf::Value inventory = boost::get<Sqf::Parameters>(params.at(5));
 	Sqf::Value hitPoints = boost::get<Sqf::Parameters>(params.at(6));
@@ -427,7 +427,7 @@ Sqf::Value HiveExtApp::loadPlayer( Sqf::Parameters params )
 
 Sqf::Value HiveExtApp::loadCharacterDetails( Sqf::Parameters params )
 {
-	int characterId = Sqf::GetIntAny(params.at(0));
+	Int64 characterId = Sqf::GetIntAny(params.at(0));
 	
 	return _charData->fetchCharacterDetails(characterId);
 }
@@ -436,7 +436,7 @@ Sqf::Value HiveExtApp::loadTraderDetails( Sqf::Parameters params )
 {
 	if (_srvObjects.empty())
 	{
-		int characterId = Sqf::GetIntAny(params.at(0));
+		Int64 characterId = Sqf::GetIntAny(params.at(0));
 
 		_objData->populateTraderObjects(characterId, _srvObjects);
 
@@ -464,7 +464,7 @@ Sqf::Value HiveExtApp::tradeObject( Sqf::Parameters params )
 Sqf::Value HiveExtApp::recordCharacterLogin( Sqf::Parameters params )
 {
 	string playerId = Sqf::GetStringAny(params.at(0));
-	int characterId = Sqf::GetIntAny(params.at(1));
+	Int64 characterId = Sqf::GetIntAny(params.at(1));
 	int action = Sqf::GetIntAny(params.at(2));
 
 	return ReturnBooleanStatus(_charData->recordLogin(playerId,characterId,action));
@@ -472,7 +472,7 @@ Sqf::Value HiveExtApp::recordCharacterLogin( Sqf::Parameters params )
 
 Sqf::Value HiveExtApp::playerUpdate( Sqf::Parameters params )
 {
-	int characterId = Sqf::GetIntAny(params.at(0));
+	Int64 characterId = Sqf::GetIntAny(params.at(0));
 	CharDataSource::FieldsType fields;
 
 	try
@@ -594,7 +594,7 @@ Sqf::Value HiveExtApp::playerUpdate( Sqf::Parameters params )
 
 Sqf::Value HiveExtApp::playerInit( Sqf::Parameters params )
 {
-	int characterId = Sqf::GetIntAny(params.at(0));
+	Int64 characterId = Sqf::GetIntAny(params.at(0));
 	Sqf::Value inventory = boost::get<Sqf::Parameters>(params.at(1));
 	Sqf::Value backpack = boost::get<Sqf::Parameters>(params.at(2));
 
@@ -603,7 +603,7 @@ Sqf::Value HiveExtApp::playerInit( Sqf::Parameters params )
 
 Sqf::Value HiveExtApp::playerDeath( Sqf::Parameters params )
 {
-	int characterId = Sqf::GetIntAny(params.at(0));
+	Int64 characterId = Sqf::GetIntAny(params.at(0));
 	int duration = static_cast<int>(Sqf::GetDouble(params.at(1)));
 	int infected = Sqf::GetIntAny(params.at(2));
